Fix to add msad.ms.com with IP info to dsdb.
Addresses-Issue: Jira/AQUILONAQD-280
The hosts were mistakenly skipped by a previous commit to support
windows hosts with dynamic IPs.

As a side effect, if the aq client is bypassed, it is now
possible to add a host without an IP address.  Hopefully this
will all be reworked by the DNS rewrite anyway.