update_model: Added command for fixing models.
Addresses-Issue: Jira/AQUILONAQD-281
Allows for fixes to machine specs (and auto-updating machine plenaries)
without the need for database surgery and scheduling turnovers.

There is no support for changing the machine type for the model at
this time.  There is also no support for updating existing machines
if the disk type (nas/local) changes.