Updated comments about the use of nyaqd1.ms.com in tests.
A test now relies on nyaqd1.ms.com being resolvable through DNS
which will not work in outside locations.  Some ideas for future
enhancements were left in the comments.

Also switched to using socket.gethostbyname() as per Gabor's
suggestion - no need to hard code the VIP.  The server is already
relying on DNS resolution so having it in the test doesn't hurt.