pxeswitch: Allow configure option to stack.
The configure option can now be given simultaneously with any of the
target options (install, localboot, firmware, blindbuild).

The status option must still be run alone.

The aii-installfe command would technically allow any of the options
to stack, and the aqd implementation is the same way.  The restriction
is only imposed client-side.