Fix memory leak
logging.Handler.removeHandler() does not remove the handler from
logging._handlers and logging._handlerList. Since our handler contains a
RequestStatus which in turn retains a copy of all log messages for the
given command, this means that all log messages ever produced are kept
in memory forever.

This behavior was introduced in commit 31463 to Python, see
http://svn.python.org/view/python/trunk/Lib/logging/__init__.py?view=diff&r1=31462&r2=31463
around line 1010.