get: Move git clone logic to the client side.
This allows the client to do better error handling than what we could
easily cram into the one-line shell script previously being pushed back
by the broker.

The client now expects CSV data of the git URL, the sandbox name, and
the user's sandbox directory.  With this info it attempts to clone the
sandbox.