Fix searching by branch in search_host and search_esx_cluster.
The intent was to be able to use the --branch option in place of
either --domain or --sandbox (without a user).  The sandbox logic was
broken in that specifying --branch sandboxname was taken as specifying
a null author.  This would never match any host.

The search has been loosened to only match on a sandbox author if one
was given using the --sandbox option.