Move host raw formatting over to mako.
This gives an example of how to redirect back into the formatting
system.  The objects being formatted may or may not be using mako
templates - it all just works.

The one catch is that the result of the redirect may be a multi-
line output.  In order to indent it properly the redirect needs to
be run through a custom filter that I called 'shift'.  The filter
just indents all lines by two spaces.  In practice I doubt this
will be a problem as future usage would just cut and paste this
example.