update_esx_cluster: Keep host personalities in line with the cluster.
Addresses-Issue: Jira/AQUILONAQD-300
The previous behavior was to error out if the cluster had hosts in it.
This has become extremely impractical.  The use cases have shown that
what we want to do is simply adjust the hosts to the new personality
right along with the cluster.

This solution is much simpler than I had planned.  The old code
did not run the chooser for cluster, so this update does not add any
chooser functionality.  Any existing host templates will continue to
point at the old personality until a make_cluster or a reconfigure is
run, at which point the Chooser will set up any needed bindings and
write an updated template.