Split boolean options to "boolean" and "flag"
If "foo" is "boolean", then the client will accept both --foo and
--nofoo. If "foo" is "flag", then only --foo is accepted. This finally
enables the client to pass false value for boolean options to the
server.

The default value of options can be set in input.xml, and is displayed
as part of the help.

Note that conflict handling is a little confusing: if --foo and --bar
conflicts, then "--foo --nobar", "--nofoo --bar" and "--nofoo --nobar"
will also conflict, which may or may not make sense.