rework the change status to allow for cluster status
This change adds a buildstatus column onto the cluster class
and implements barriers to promoting hosts to ready based on
the cluster buildstatus. If you ask to promote a host to ready
when it's a member of an unready cluster, then the host will
instead get moved into the "almostready" state. When the cluster
buildstatus is promoted to ready, then all the almostready hosts
will get promoted up to the real "ready" state. As part of adding
the new states, I've added transition graphs into the definition
of status and validate that hosts/clusters are only moving to
valid states according to the graph. This means that there
are numerous incompatabilities with previous processes:
1) clusters have a build status and they must be signed off
2) hosts have a new status (almostready) that people may not
   expect (although it will be presented still as "build" in ESP
   which is what we're after).
3) in order to rebuild a host/cluster, there are new states
   "rebuild" and "reinstall" that must be used (it is illegal to
   move a cluster/host back to the build state). This particular
   change requires some modifications to the installack.cgi script
   for vmhosts. This also requires a change to the datawarehouse
   script/view to present "rebuild"/"reinstall" in the same way
   as if the host were "ready" - i.e. present the business
   function as unchanged during a rebuild, instead of dropping back
   to the MIS state of "build".
I've also added in a "decoed" state while I was there, but it's not
used right at this moment in time...