add_address: Backported CLI for this and del_address.
Addresses-Issue: Jira/AQUILONAQD-307
This is a completely independent implementation of the changes pending
in the DNS revamp because none of the schema infrastructure from that
branch is in place here.  The records created are in a new FutureARecord
subclass of System that should map to the coming ARecord class.

I tweaked some of the input.xml parameters (URL paths, optional
parameters) that should probably be carried forward to the upcoming
work.  Otherwise the actual implementation can be tossed.