cluster: Renamed bind_esx_cluster --hostname.
Also makes the corresponding change for unbind/uncluster.  The old
commands are still around and issue a deprecation warning.

The only difference is that there is no rebind equivalent - the
cluster command does a straight rebind.  That is, if a host is already
in a cluster, it swaps it to the new cluster while issueing a warning.