poll_tor_switch: Fix for polling vlan information for a rack.
This change wraps the vlan2net operation in a try/except block.  If
all switches fail, the command fails.  Otherwise the failures are just
sent as warnings to a client.

This should smooth the 10gig rollout as running vlan2net against a
bor switch (management interfaces) will fail.  The long term fix is to
get some extra information into aqdb about whether a switch is a tor,
bor, or aggregate, and use that when polling.