Updated DSDB interaction as per the wiki.
If the current documentation is not clear, please update it based on
these examples.

Interaction with the database is first because it's the easiest to roll
back.  This code has explicit calls to flush which should be unnecessary
while configured for autoflush but help to emphasize the ordering.

After the information has been pushed to the database (short of a
commit), then the code calls out to external programs.  If these fail,
the transaction rolls back nicely.

Reviewed-by: Wesley Hertlein <Wesley.Hertlein@morganstanley.com>