runtests: Add a --mirror option.
This rsyncs the srcdir (everything, including .git) over to the directory
specified in the config for unittest/mirrordir.  Afterwards, runtests.py
execs the copy of itself.

The idea is to allow work to continue on other branches (or even on the
same branch) while running against a snapshot or the source.

This might eventually be the default behavior. :)