fixed the templates for interfaces in the face of aliases
Previously, the code was trying to create aliases named eth0:1.
That doesn't work - the PAN templates (a) don't allow names
like that as keys within nlists and (b) doesn't work with
the linux networking component that expects it to look like
an aliases nlist dangling off the "eth0" key. So, I've
restructured all the interface code, using a new helper class
within panutils which converts python data structures into
PAN code.

This code still doesn't do quite the right thing for network
aliases: the ordering of address-assignments is not deterministic
and therefore this may down an interface and recreate a new interface
if it sees the ordering change. That's bad. To be fixed later.

Reviewed-By: Wesley Hertlein <Wesley.Hertlein@morganstanley.com>