Always create an interface if there is a primary name
There is special logic in get_or_create_interface() so "add interface"
still succeeds if it is called with the name of the auto-created
interface. There is also special logic for allowing the removal of the
auto-created interface and migrating the primary IP if the real
interface is added later.

Reviewed-by: Wesley Hertlein <Wesley.Hertlein@morganstanley.com>