show_request: Fix where newlines are added to log records.
The newlines should always have been added after each log record by
the server.  However, the original implementation had them added by
the client.

This worked as long as the client got the records one at a time,
which it generally did.  However, if there is a proxy in line
this may not be true.  Fix by having the client just write whatever
it gets and having the server insert the newlines in the correct spots.

Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>