When the knc process fails to initialize, provide a more useful error
The error message generation was blowing up because it was assuming
it could read the error from the process, but if the process fails to
start, that won't work. It now gives a "stub" message if the knc
process won't start at all, which is a little better than a stacktrace.

Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>