Modify the kerberos service name from "--aquser" to "--aqservice"
The option "aquser" doesn't really make sense in terms of what
it's doing: describing the service principal that should be used
to connect. I've therefore changed it to aqservice. The keytab
I've made explicit in the config file, instead of hardcoded/assumed
which now allows you to (for example) have a different keytab
to the username by which you're running. For example, you could
point at the host keytab when running as root (not that you'd
want to, but it's a quick and easy test of the system)

Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>