search_host: Add an option to search for hosts using a share.
Specifically, search for a host with a cluster with machines with
disks on a given nas_disk_share.  This was requested by Ops to
figure out which servers might be affected in a share outage.

The option is rather poorly named as member_cluster_share.  I
wanted a name that wouldn't clash with the options that already
exist.

Ideally this information would be in the data warehouse and then
they could slice and dice as needed.  This is an interim solution.

Reviewed-By: Scott DAquila <daqscott@ms.com>