Use column references rather than simple names in order_by clauses
This avoids a potential problem when a query involves two tables, both
having the same column name that is used in the order_by, therefore the
generated SQL will be ambiguous. With column references, SQLAlchemy
qualifies the column name with the right (probably aliased) table name,
so this problem will not occur.

Reviewed-By: Wesley Hertlein <Wesley.Hertlein@morganstanley.com>