Fix query generated by search_host
Addresses-Issue: Jira/AQUILONAQD-346

When the same table is joined multiple times with different filters,
SQLAlchemy doesn't generate aliased table references in the filter
conditions, which makes the column names ambiguous. Re-arrange the code
to avoid the need of multiple joins, and add some hints to reduce the
number of queries.

Reviewed-By: Wesley Hertlein <Wesley.Hertlein@morganstanley.com>