Update/access service bindings through relations
Creating/deleting an object directly does not update relations that are
already loaded. Calling the appropriate method of the relation however
allows SQLA to keep the relation up to date without reloading it from
the database. There is a catch though: when deleting an object that is
part of multiple relations, only one of those relations can be updated
this way, and the others have to be invalidated explicitely using
session.expire().

Reviewed-By: Wesley Hertlein <Wesley.Hertlein@morganstanley.com>