Add support for creating VLAN interfaces
When adding switches without interface information but with an IP
address, a dummy interface is created where the IP address can be bound.
As a side effect of this change, these dummy interfaces are
automatically removed when a real interface is added later, and the
primary IP is automatically re-assigned to the real interface.

Reviewed-By: Wesley Hertlein <Wesley.Hertlein@morganstanley.com>