add in protobuf support for locations with timezones
In order to be able to improve the datawarehouse
with geographic information, I've exposed the location
information via protobufs and CSV. Importantly for the
datawarehouse, the output includes the timezone
information so I've used this opportunity to clean
up the timezones used by cities in AQDB. Previously,
timezones were lowercased (but TZ is case sensitive)
there was no way to edit them against a city and the
default value was something that would break any
consumer. I've changed the default to be UTC. The
add_city and update_city commands now provide
timezone support and expose the data via plenary templates.
By providing the plenary template, this also avoids the need
for people to create new site templates when adding in a
new datacenter.

Reviewed-By: Wesley Hertlein <Wesley.Hertlein@morganstanley.com>