move the "aq flush" test much earlier to get templates in-sync
I've moved the flush command to happen earlier in the tests.
This means that not only does the flush command get tested, but it
also does the neccessary task of ensuring the plenary templates
are up-to-date for the data imported from the unittest.data
bootstrap. Without the flush, later "aq reconfigure" and "make
aquilon" tests are not using the full range of plenary templates
(e.g. the city templates). This was caused no previous issues because
the location templates are only included if present and all other
templates were constructed on the fly.

Note that the position of the flush test is important. We
require it to be before any compiles happen (to make sure we
have all the plenaries, even for imported data), but after we've
made some services and machines (to make sure we get sufficient
test coverage of the flush command iterating over all the
different functionality).

Reviewed-By: Wesley Hertlein <Wesley.Hertlein@morganstanley.com>