Stop labelling non-addressed interfaces as "static" IP
Previously, an interface's bootproto was either "dhcp" or "static".
However, "static" bootproto implies to the system that there should
be configuration for IP address, netmask, etc.
We sometimes have interfaces without an IP address but labelled as
"static" bootproto and that's causing the
network configuration component to emit errors and cause the host
to fail signoff. An interface without any IP address should be
given the bootproto of "none".