config: Add a fallback for setting the default user variable.
If the USER environment variable does not exist, fall back to
checking the password database.  I don't remember why I didn't
use this method in the first place... maybe ease of testing
alternate configurations?

As a config option, this value is (still) not meant in any way,
shape, or form to be used for security.

Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>