Register manager/auxiliary addresses in DSDB even if there is no host
The management address should be added to DSDB even if the machine has
no host entry (and therefore no primary name). Do not use the
-primary_host_name flag when calling "dsdb add_host" for the management
address even if there is a host entry.

When checking if DSDB needs to be updated, the change of the primary
name means old auxiliary addresses will be removed and re-added with the
correct -primary_host_name flag.