actual tests of using location plenaries
The last addition of location plenary templates (cities)
didn't really test the templates, since all the locations in
question were covered by bootstrap data provided before the
test starts. I've created a new set of locations which is
constructed via broker commands instead of bootstrapped. The
intent (although it's not complete yet - only up to the city
level) is that the stack of commands would be system tested
rather than using bootstrap data. To this end, there's a new
building called "cards" (as in house of cards... seemed like
a good idea at the time). This building is in the "ex"
(example) city and uses the DNS domain of cards.example.com.
A machine is placed into a rack in this building; a host
is configured on that machine; the host is compiled and
then finally the building and city (and all associated
networks and service mappings) are torn down.

tests show: Ran 1804 tests in 2310.421s

Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>