Handle host templates that require cluster information.
Addresses-Issue: Jira/AQUILON-300
This adds a cluster_required flag to archetype.  Trying to write a
host plenary for a flagged archetype will fail with an IncompleteError
if the host is not clustered.

When running uncluster on such a host any existing templates will be
cleaned up.

Includes an enhancement to update_archetype to handle the new flag,
and a tweak to the existing behavior.

Reviewed-By: Scott DAquila <daqscott@ms.com>