Do not write host plenary templates for fake hardware.
This is a proposed compromise solution for the production change
awhile back to mark the aurora archetype as compilable.  There
was already a check in the machine plenary to skip writing the
machine plenary if the model is aurora_node - this is a fake model
only used by add_aurora_host.  The logic was copied for the host
plenary templates.

This should allow us to create and compile profiles for aurora
hosts with real hardware while not breaking dummy entries.  At
some point in the future the aurora_node model should be
removed completely anyway.

Reviewed-By: Scott DAquila <daqscott@ms.com>