Add a readonly mode to the broker.
This is not strictly a readonly mode - new users will still be recorded
into the database with role 'nobody', and we could potentially add
auditing and logging hooks.

When the broker is in this mode, it will refuse to execute commands
that do not set up a readonly session into the database.

Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>