Create aqd and aqd_readonly symlinks for the twistd script.
When the 'aqd' symlink is used the broker is forced into mode
readwrite.  When the 'aqd_readonly' symlink is used the mode
becomes readonly.  Use of the symlinks overrides whatever is
set in the config file.  If the symlink name conflicts with the
config file setting a log message is generated.

Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>