compile: Add debug output for panc invocations on request.
Addresses-Issue: Jira/AQUILONAQD-385

In addition to adding options for enabling the panc debug output,
this commit also adds an option to clean out the dependency files
generated by pan, effectively forcing a rebuild.

It is tremendously useful for regression and timing tests, and might be
needed in production if swapping panc versions that change dependency
formats (such as downgrading from 8.4.x to 8.2.x).  I don't really want
to advertise it for general use... if everything is working as it
should, the option should not be necessary.

Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>