support for a wider range of clusters
Add in generic commands for add, show, update, and del cluster.
When you specify an archetype that will associate the cluster with a
specific subclass with its own data. That mapping between archetype
and database tables is maintained within the archetype (cluster_type).
I've also added in a description field to the archetype which is used
to present a human readable form of the archetype during raw output.

The "storage cluster" is still a bit sparse.

I've moved some data that was originally specific for esx clusters into
more generic models and commands. For example, down_hosts_threshold,
which I've also modified to allow for it to be either an absolute number
or a percentage (which makes more sense for grid clusters).

Personalities are either host or cluster, and changes throughout
relevant commands to cope with this. Notably the "cluster"/"uncluster"
command now does not change the host personality to match the cluster
personality, but does allow for changing the personality on entry/exit
of the cluster to allow for moving the host into a personality that can
cope with the new state (actually, if the existing personality is
marked as requires_cluster, it should require that... TBD).

The old ESX "vmhost" archetype is split into two - the ESX hosts keep
vmhost and the clusters are now in the esx_cluster archetype.

The "cluster_required" flag I've moved out of archetype and
into personality (so that I can move hosts in and out of clusters
without changing the archetype of the host).

I've added a "is_cluster" property of archetypes to allow
for cluster-specific archetypes.

Previously, there were some checks about using a valid
personality for a host (!is_cluster personality). Since I dropped
the idea of changing archetype, and it's now the archetype that
says if it's a cluster or host style, the check is irrelevant.
However, checking that the personality on exit is not
cluster_required is still relevant.

All clusters may be defined at a campus level.

Adds protobuf support for clusters.

Cluster aligned services now for any type of cluster, not just esx.

Add in commands for manipulating allowed personalities of clusters.

Reviewed-By: Wesley Hertlein <Wesley.Hertlein@morganstanley.com>