DHCP config generation fixes for Zebra and bonding
In case of Zebra, the problem was that the archetype name was set only
when emitting the primary IP, but the host wants to do PXE using the
transit IP. This caused aqdhcpd to ignore the transit address. Now the
archetype is set for all addresses of a host.

In case of bonding, the problem was that the physical interface had the
MAC address, but not the IP address, and the logical address vice versa,
so aqdhcpd ignored both. Now we output a record for every MAC address
a given IP address may show up with.

Reviewed-By: Wesley Hertlein <Wesley.Hertlein@morganstanley.com>