Don't populate MAC addresses for non-bootable public interfaces
Sending two host messages when only the MAC address differs causes
aqdhcpd to create two "host {...}" sections using the same hostname in
case of bonding interfaces, and that is not a valid configuration. If we
want to fix this properly, then aqdhcpd should be taught to either
ignore the address it does not want, or emit two "hardware address ..."
statements inside the same "host {...}" section. Alternatively the
protocol could be updated to allow including the extra information
some other way.

Reviewed-By: Wesley Hertlein <Wesley.Hertlein@morganstanley.com>