Fix for an aq compile test of proper dependency checking in panc.
The results varied based on randomness in the chooser.  The test
has been updated to remove the hard coded number of hosts and
calculate it on the fly.

Unfortunately there is a bug in the panc task dependency checker
that is forcing a recompile of all cluster profiles.  The test
takes that into account and will need to be fixed when the panc
bug is fixed.