automac: Fixed algorithm to only search holes if needed.
The database search for existing mac addresses was too wide.  Sadly,
when I wrote the initial search, I did not know or remember that less
than, greater than, and between comparisons worked on strings in SQL.
Fixed the search to only return mac addresses in the range we care
about.  Now the algorithm to search for holes will only kick in if the
last mac in the range has been taken.

Also verified that BETWEEN for both sqlite and oracle follow the spec
that the keyword be inclusive for both data points.

Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>