autopg load balances across portgroups
Changed the way autopg behaves when adding an interface for a VM.
Prior to this commit interfaces are added to a vlan until filled
before moving onto the next.  Post commit, the interfaces are now
"load balanced" across the vlans/port groups.

To load balance vlan.guestcount is repeatedly called for all user
vlans and selecting the least loaded vlan.  For a Vulcan rack of 850
VMs, this increases the DB calls for add interface from ~1500 to ~3400

As part of the commit the unit tests were also updated to verify this
change.

Reviewed-By: Wesley Hertlein <Wesley.Hertlein@morganstanley.com>