Add config option to advertise gzip'd profiles as simple .xml files.
Addresses-Issue: Jira/AQUILON-365
The new option only does anything when profiles are being gzip'd.  When
enabled, instead of listing the files as .xml.gz they will simply be
listed as .xml.  Also, the profiles-info.xml file will actually be
written as profiles-info.xml.gz.

This relies on webserver magic to properly retrieve the compressed
files (decompressing if the client does not support a gzip encoding)
when a client requests the .xml.

The regression tests were run four times - once for all the
combinations of the two config values (this one and the old
gzip_output option).

Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>