aq: Add no-half-close to the knc options.
If the server end of the knc connection dies, knc will close stdout on
the client side.  However, knc stays running.  When the aq client tries
to read stderr, it blocks.

Setting this option results in knc shutting down cleanly when the pipe
on the server side is no longer valid, which is the desired behavior.

Addresses-Issue: Jira/AQUILONAQD-295

Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>