Added tools and README for use outside MS systems.
The README documents how to get a SLES 6 system installed and have
the broker start in a development mode.  It is something of a work
in progress.

Of note is a new build script called bootstrap_env.py.  This uses
the virtualenv module (not available in MS) to help create an
isolated environment with all the necessary dependencies to run.

It has two hacks.  First, it assumes the use of the system protobuf
libraries (which need to be installed).  I couldn't get that working
with ez_setup.  Second, it uses some empty module implementations to
provide stubs for ms.* modules.

Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>