Rename the aquilon.server module to aquilon.worker, part 1.
In preparation for future work on the multiprocess branch (where there
are multiple modules that make up the server), renaming the
aquilon.server module as aquilon.worker.

Note that this commit only renames the directory.  This is done so that
there is no ambiguity for git in following file history.  The next
commit will fix all the imports both internal and external to the
module.

Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>