Rename the aquilon.server module to aquilon.worker, part 2.
In preparation for future work on the multiprocess branch (where there
are multiple modules that make up the server), renaming the
aquilon.server module as aquilon.worker.

Note that this commit relies on the files already being renamed.
It fixes all the imports both internal and external to the module.

Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>