Fix network location refresh
We always have to load all networks from AQDB to be able to detect
location mismatch, even if the --building option was passed to
refresh_network. However we have to "forget" the networks that are not
inside the requested building before we decide that they are missing
from QIP and therefore should be deleted.

Reviewed-By: Wesley Hertlein <Wesley.Hertlein@morganstanley.com>