Test server notifications
Send the notifications to the local host, where there is little chance
they cause any harm. Unfortunately we can't really test client
notifications, because we do not want to send garbage to random IP
addresses.

In the future we could also launch a listener process to test if the
notifications in fact arrive.

Reviewed-By: Wesley Hertlein <Wesley.Hertlein@morganstanley.com>