fixed up the cluster tests
Added in another version of esx in order to do some reconfigure tests,
minor fixes around the tests and a new testcase for building a db
standalone/cluster.

In the test to change the personality of the cluster members, we were
previously switching the personality to something that didn't work.
I've now tried to switch to a personality that actually compiles.

Reviewed-By: Wesley Hertlein <Wesley.Hertlein@morganstanley.com>