add in resources (intervention, filesystem, app)
Resources are ephemeral "things" attached to a host or a cluster.
This resource adds in the base class of resources with some
example subclasses: filesystem, application and intervention.

The filesystem resource is the most fleshed out example, but
interventions and applications also have some content.

The template name for a resource includes the holder:
  /resource/filesystem/host/host.example.com/fs1/config.tpl
  /resource/filesystem/cluster/grid1/fs1/config.tpl

The host object template includes any cluster resources
directly (under /system/cluster/resources instead of
/system/resources) in case the host wants them.

Includes protobuf support.

Reviewed-By: Wesley Hertlein <Wesley.Hertlein@morganstanley.com>