POC: dense location relations
Instead of storing a single link to a parent location, store links to
all ancestors. The distance is maintained in the link, so it is still
possible to identify direct parents.

The only casulty is "update rack" as reparenting will now be more
complicated. The missing logic would be similar to what
Location.__init__ does when it receives a "parent" option.

Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>