cloneprod: Add an option to clone only the schema from prod.
This is useful for doing an upgrade test, although it wouldn't test
data migration.  The idea is to do the schemaonly exp/imp from prod,
run the upgrade script(s), run build_db with the populate option, and
then try running the unit tests.

It's still useful to run in normal mode and do a "full" upgrade test,
poking at a dev environment afterwards to see if everything looks
right.

While I was in here, I added some documentation to the help text
around how I'm using the tool.

Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>