working upgrade sql scraps for this branch
The existing ESX clusters need to have their personality switched to a
cluster personality, so I've done that in the data.sql part of the
upgrade.  I've created two personalities 'vulcan-v1-1g-prod' and
'vulcan-v1-10g-prod'.  Any existing cluster that had a 10g personality
will be converted to the new 10g personality.  All other clusters will
be converted to the 1g personality.

Reviewed-By: Wesley Hertlein <Wesley.Hertlein@morganstanley.com>