change the default os from 4.0.1 to 5.0.1
Testing new features (e.g. vcs) which are being added into
current OS's is easier if the test suite is using a current OS.

The tests (and one bit of broker code in add_host) were using
linux 4.0.1 as the default OS. That OS has been deprecated
for some time (4.0.2 should be used at least) and we're currently
adding in new features into 5.x and 6.x. So, I've upped the default
to 5.0.1.  When testing switching OS, I'm using 5.0 as the
alternative.

Reviewed-By: Wesley Hertlein <Wesley.Hertlein@morganstanley.com>