Display the network environment for interfaces
The network environment is a property of AddressAssignment, but we do
not allow mixing multiple network environments on the same interface, so
displaying it as a property of Interface makes sense.

At the same time remove the "Attached to:" line, because it's
superfluous: interfaces are displayed only as part of a hardware entity,
never alone.

Reviewed-By: Wesley Hertlein <Wesley.Hertlein@morganstanley.com>