deploy: Close deploy to sandbox loophole.
The target option specifically looks for a Domain.  There is a bit
of magic that lets the user specify a target as a tracking domian,
and deploy will just pretend the user specified the tracked domain
and give a warning.

If the tracked branch was actually a sandbox, the magic still
happened but then would give an internal error later on when the
code accesses an attribute that only Domains have.

This change tweaks the check to make sure that the target is a
Domain before letting the swap magic kick in.  The error message
that the user gets will be a little confusing, but it is better
than the internal error.

Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>