search_audit: Fix broken behavior when no timezone given.
The documentation indicates that it should be possible to give a date
without a timezone to the before and after options.  The timezone for
this date was not defined.

Fixed the command to allow a date/time without a timezone.  UTC is
used as the default.  Added this information to the documentation
and added a test.

Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>