search_audit: Fix raw output to have a consistent time display.
On Oracle, only seconds are stored.  On sqlite, microseconds are
stored.  The str() formatter for datetime behaves differently
depending on whether microseconds == 0.

While moving to strftime to give an explicit format, had to change
the offset format.  The strftime method only offers %z, which does
not include the ':' that str() and isoformat() use.

Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>