Fix escaping when generating pan templates
Unfortunately the pan escape() function does not always leave the string
unchanged even if panc would accept it as a valid nlist key. Also,
calling unescape() on a string which did not go through escape() is not
safe if the string contains an underscore followed by digits. Therefore
the templates must use escape() if and only if the receiving side calls
unescape(); it is not possible to escape() strings automatically.

Using escape() too much on the other hand reduces the readability of the
templates, so the new PanEscape object still tries to avoid emitting
escape() if there is no danger of unescape() misinterpreting the value.

Addresses-Issue: Jira/AQUILON-399

Reviewed-By: Wesley Hertlein <Wesley.Hertlein@morganstanley.com>