Defer looking up the user after creating the audit record
Adding the command to the audit tables means a COMMIT, which in turn
invalidates any objects loaded before. So if the user is looked up
before adding the audit record, then the query must be re-issued when
the authorization checks access the user object. The solution is to move
the user lookup after the creation of the audit record.

Reviewed-By: Zoltan Sebestyen <Zoltan.Sebestyen@morganstanley.com>