Fix query option use in HardwareEntity.get_unique()
We need to run the query against the appropriate class in order to be
able to use options for attributes not defined in the base class.  Since
we still want to detect when the name already exists but belongs to a
different hardware type, we must do two queries: one against the child
class to make sure Query.options() work correctly, and one against
HardwareEntity if the first one failed.

As a side effect, make sure that the primary name is not loaded eagerly
when we already know it.

Reviewed-By: Zoltan Sebestyen <Zoltan.Sebestyen@morganstanley.com>