Relax eager loading on the backref when subqueryload() is used
If A.b is a relation to B that is eager-loaded by default, then using
subqueryload() from the other direction (i.e.,
session.query(B).options(subqueryload(B.a))) will cause B to be loaded
twice. To avoid this, we need to explicitly cancel the eager-loading by
using lazyload().

Eventually, we should get rid of doing eager-loading directly in the
model because it's hard to get such policy right there, but that is a
bigger change.

Change-Id: I75609e110d12aa0b46dc3a8a54ad8181f95dea64
Reviewed-by: Zoltan Sebestyen <Zoltan.Sebestyen@morganstanley.com>