Add parent locations to machine's location field
Description: At the moment we only get the most specific location object
in the host message although the message spec allows it to to a list of
all its ancestors. This diff causes the names and types of the ancestors
to get added to the location protobuf.

The unit test testverifyshowhostproto was also modified to check that
machine.location is correct and it has exactly 7 parents.

Test Plan: Ran unit tests and ran in ipython:
In [47]: [(l.name, l.location_type) for l in
host.machine.location.parents]
Out[47]:
[(u'ms', u'company'),
 (u'ny', u'hub'),
 (u'na', u'continent'),
 (u'us', u'country'),
 (u'ny', u'campus'),
 (u'ny', u'city'),
 (u'np', u'building'),
 (u'np-lab1', u'room')]

Addresses-Issue: Jira/AQUILON-420

Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>