show_active_commands: Fix for zombie listings.
We have seen finished commands listed in the show_active_commands
output in production for a few months now and I never tracked down
the root cause.  I was inspired earlier today when I noticed that
all the listings involved bad arguments (like a poorly formatted
mac address).

As it turns out, the broker parameter check happened before the
try/finally block that cleans up the in-memory status on the broker.

I had a choice of location for it - either immediately after the
try block opens or later on after an audit record is created.  I
decided to leave it after the audit record gets created.  There is
no real need to log the error condition, but it seems best from a
consistency stand point to do so.

Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>