poll_switch: Execute CheckNet from a local server.
When running CheckNet in NY against a switch across the world, it
can take a very long time.  This change runs the command via ssh
on a server close to the switch's.

The server is determined by looking for the a configurable service
(defaults to poll_helper) mapped to the location of the switch, and
then walking the location tree. If no server is found, CheckNet is
run on the broker.

This is something of a hack (since aqd might not be the right tool
for the job as opposed to an aii tool) to get the work done quickly.

Anyone with a custom unittest configuration should add this line in
the broker section:
poll_ssh = %(srcdir)s/tests/fakebin/fake_ssh

Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>